@@ -1,44 +1,47 @@
 package Model;
 
-import java.util.ArrayList;
-import java.util.Date;
-import java.util.HashMap;
+
+import java.time.LocalDateTime;
+import java.util.*;
 
 public class Off extends SaveAble {
-    private static ArrayList<Off> allOffs;
-    private static HashMap<String, Off> offs;
+    private static ArrayList<Off> allOffs = new ArrayList<>();
     private String offBarcode;
-    private ArrayList<Product> products;
+    private ArrayList<String> products;
     private enum offStatus {
         MAKING, EDITING, APPROVED
     }
 
-    private Date startDate;
-    private Date endDate;
-    private double offAmount;
+    private LocalDateTime startDate;
+    private LocalDateTime endDate;
+    private int offAmount;
 
-    public Off(String offBarcode, Date startDate, Date endDate, double offAmount) {
-        this.offBarcode = offBarcode;
-        this.products = new ArrayList<Product>();
+    public Off(LocalDateTime startDate, ArrayList<String> products, LocalDateTime endDate, int offAmount) {
+        this.offBarcode = givenUsingPlainJava_whenGeneratingRandomStringUnbounded_thenCorrect();
+        this.products = products;
         this.startDate = startDate;
         this.endDate = endDate;
         this.offAmount = offAmount;
+        for (String product: products) {
+            Product.getProductWithBarcode(product).offTheCost(Product.getProductWithBarcode(product).getCost() * offAmount / 100);
+        }
         allOffs.add(this);
+        SaveAndLoad.getSaveAndLoad().writeJSON(allOffs, ArrayList.class, "allOffs");
     }
 
     public String getOffBarcode() {
         return offBarcode;
     }
 
-    public Date getStartDate() {
+    public LocalDateTime getStartDate() {
         return startDate;
     }
 
-    public Date getEndDate() {
+    public LocalDateTime getEndDate() {
         return endDate;
     }
 
-    public double getOffAmount() {
+    public int getOffAmount() {
         return offAmount;
     }
 
@@ -58,7 +61,51 @@ public class Off extends SaveAble {
         return offBarcode;
     }
 
-    public ArrayList<Product> getProducts() {
+    public ArrayList<String> getProducts() {
         return products;
+    }
+
+    public static Off getOffByBarcode(String name) {
+        for (Off off: allOffs) {
+            if (off.getOffBarcode().equalsIgnoreCase(name)) {
+                return off;
+            }
+        }
+        return null;
+    }
+
+    public static ArrayList<Off> getAllOffs() {
+        return allOffs;
+    }
+
+    public String givenUsingPlainJava_whenGeneratingRandomStringUnbounded_thenCorrect() {
+        int leftLimit = 97; // letter 'a'
+        int rightLimit = 122; // letter 'z'
+        int targetStringLength = 10;
+        Random random = new Random();
+        StringBuilder buffer = new StringBuilder(targetStringLength);
+        for (int i = 0; i < targetStringLength; i++) {
+            int randomLimitedInt = leftLimit + (int)
+                    (random.nextFloat() * (rightLimit - leftLimit + 1));
+            buffer.append((char) randomLimitedInt);
+        }
+        String generatedString = buffer.toString();
+
+        return generatedString;
+    }
+
+
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        Off off = (Off) o;
+        return offBarcode.equalsIgnoreCase(off.offBarcode);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(offBarcode) + 100;
     }
 }
