@@ -6,15 +6,15 @@ import java.util.ArrayList;
 public class Comment extends SaveAble {
     private String commentBarcode;
     private static ArrayList<Comment> allComments;
-    private Account commentingAccount;
-    private Product product;
+    private String commentingAccount;
+    private String product;
     private String commentText;
     private enum commentStatus {
         PROCESS, APPROVED, REJECTED
     }
     private boolean boughtTheProductOrNot;
 
-    public Comment(Account commentingAccount, Product product, String commentText, boolean boughtTheProductOrNot) {
+    public Comment(String commentingAccount, String product, String commentText, boolean boughtTheProductOrNot) {
         this.commentingAccount = commentingAccount;
         this.product = product;
         this.commentText = commentText;
@@ -26,7 +26,7 @@ public class Comment extends SaveAble {
         return commentBarcode;
     }
 
-    public boolean hasBoughtTheProductOrNot (Customer customer) {
+    public boolean hasBoughtTheProductOrNot (String customer) {
         return false;
     }
 
