@@ -1,32 +1,49 @@
 package Model;
 
+import Control.Controller;
+
+
+import java.time.LocalDateTime;
 import java.util.ArrayList;
+import java.util.HashMap;
 
 public class Customer extends Account {
     //private static ArrayList<Customer> allCustomers = new ArrayList<Customer>();
-    private ArrayList<Product> cart;
+    private HashMap<String, Integer> cart;
+    private ArrayList<String> sellersOfProductsOfTheCart;
     //private ArrayList<BuyLog> buyingHistory;
     //private ArrayList<CodedOff> offCodes;
     protected ArrayList<History> history;
-    protected ArrayList<CodedOff> offCodes;
+    protected ArrayList<String> offCodes;
+    protected ArrayList<Integer> usageOfOffCodes = new ArrayList<>();
+    private String address = "";
 
     public Customer(String userName, String passWord) {
         super(userName, passWord);
         this.offCodes = new ArrayList<>();
-        //allCustomers.add(this);
         this.history = new ArrayList<>();
+        this.cart = new HashMap<>();
+        this.sellersOfProductsOfTheCart = new ArrayList<>();
+//        allCustomers.add(this);
+        SaveAndLoad.getSaveAndLoad().writeJSON(this, Customer.class, userName);
+        for (String offCode: offCodes) {
+            usageOfOffCodes.add(0);
+        }
     }
 
     public static void newCustomer(String username, String password) {
-        Customer customer = new Customer(username, password);
-        SaveAndLoad.getSaveAndLoad().writeJSONAccount(customer);
+        new Customer(username, password);
     }
 
-    public ArrayList<Product> getCart() {
+    public void setAddress(String address) {
+        this.address = address;
+    }
+
+    public HashMap<String, Integer> getCart() {
         return cart;
     }
 
-    public ArrayList<Product> getProducts() {
+    public ArrayList<String> getProducts() {
         return null;
     }
 
@@ -35,8 +52,7 @@ public class Customer extends Account {
 //    }
 
 
-
-    public ArrayList<CodedOff> getOffCodes() {
+    public ArrayList<String> getOffCodes() {
         return offCodes;
     }
 
@@ -48,20 +64,48 @@ public class Customer extends Account {
         return null;
     }
 
-    public void buy(Product product) {
+    public void buy(String product) {
 
     }
 
-    public static void addProductToCart(Product product) {
-        //check if exiits or not and kam kon azash
+    public int addProductToCart(String product) {
+        if(Product.getProductWithBarcode(product).isExistsOrNot()){
+            Product.getProductWithBarcode(product).setAmountOfExist(Product.getProductWithBarcode(product).getAmountOfExist() - 1);
+            cart.put(product, cart.get(product) + 1); // maybe need edition
+            return 1;
+        }
+        else{
+            return 0;
+        }
     }
 
-    public boolean pay() {
-        //check if money is enough
-        if (this.getCredit() < getCartMoney()) {
+    public boolean pay(String offCode) {
+        //age chizi be Off ezafe shod bayad costesh hamoon ja kam beshe haaaaaa in ja off mohasebe nemishe va faghat codedOff ha tasir daran
+        //too saef mahsool bayad darj beshe ke kodoom seller dare ino mofrooshe va too customer ye arrayList hast ke be tartbie product haye hashMap product haye cart seller haye har product rpo ham zakhire mikone
+        int finalCost = getCartMoney();
+        if (CodedOff.getOffCodeWithName(offCode) != null && this.usageOfOffCodes.get(getOffCodeIndexForUsageTimeAddingByName(offCode))  < CodedOff.getOffCodeWithName(offCode).getUsageTime()) {
+            if ((getCartMoney() * CodedOff.getOffCodeWithName(offCode).getPercent() / 100) > CodedOff.getOffCodeWithName(offCode).getOffAmount()) {
+                finalCost -= CodedOff.getOffCodeWithName(offCode).getOffAmount();
+            } else {
+                finalCost -= (getCartMoney() * CodedOff.getOffCodeWithName(offCode).getPercent() / 100);
+            }
+            this.usageOfOffCodes.set(getOffCodeIndexForUsageTimeAddingByName(offCode), this.usageOfOffCodes.get(getOffCodeIndexForUsageTimeAddingByName(offCode)) + 1);
+        }
+        if (this.getCredit() < finalCost) {
             return false;
         } else {
-            this.credit -= getCartMoney();
+            addHistory(finalCost);
+            this.credit -= finalCost;
+            ArrayList<String> products = (ArrayList<String>) cart.keySet();
+            int i = 0;
+            for (String seller: sellersOfProductsOfTheCart) {
+                Seller sellerFromFile = (Seller) Seller.getAccountWithName(seller);
+                sellerFromFile.setCredit(Product.getProductWithBarcode(products.get(i)).getCost() * cart.get(products.get(i)));
+                SaveAndLoad.getSaveAndLoad().writeJSON(sellerFromFile, sellerFromFile.getClass(), sellerFromFile.getUserName());
+                i++;
+            }
+            cart = new HashMap<>();
+            SaveAndLoad.getSaveAndLoad().saveGenerally();
             return true;
         }
     }
@@ -69,11 +113,20 @@ public class Customer extends Account {
     public int getCartMoney() {
         int cartCost = 0;
         if (this.cart.size() != 0) {
-            for (Product product : cart) {
-                cartCost += product.getCost();
+            for (String product : cart.keySet()) {
+                cartCost += Product.getProductWithBarcode(product).getCost() * cart.get(product);
             }
         }
         return cartCost;
+    }
+
+    public int getOffCodeIndexForUsageTimeAddingByName(String name) {
+        for (int i=0; i<offCodes.size(); i++) {
+            if (Off.getOffByBarcode(offCodes.get(i)).getOffBarcode().equalsIgnoreCase(name)){
+                return i;
+            }
+        }
+        return -1;
     }
 
     @Override
@@ -89,5 +142,39 @@ public class Customer extends Account {
                 ", passWord='" + passWord + '\'' +
                 ", credit=" + credit +
                 '}';
+    }
+
+
+    public void setNumberOfProductInCart(String productInCart, int n) {
+        if (cart.keySet().contains(productInCart)) {
+            cart.replace(productInCart, cart.get(productInCart) +n);
+            if (cart.get(productInCart) == 0) {
+                cart.remove(productInCart);
+            }
+        } else {
+            System.out.println("This barcode is none of your selected products in cart");
+        }
+
+    }
+
+    public History getHistoryById(String name) {
+        for (History history: history) {
+            if (history.getLogBarcode().equalsIgnoreCase(name)) {
+                return history;
+            }
+        }
+        return null;
+    }
+
+    public void addHistory(int finalCost) {
+        LocalDateTime dateTime = LocalDateTime.now();
+        int offCost = getCartMoney() - finalCost;
+        ArrayList<String> products = new ArrayList<>(cart.keySet());
+        History historyOfPurchase = new History(dateTime, getCartMoney(), offCost, sellersOfProductsOfTheCart, products);
+        history.add(historyOfPurchase);
+    }
+
+    public void addOffCode(String offCode) {
+        this.offCodes.add(offCode);
     }
 }
