@@ -1,25 +1,31 @@
 package Model;
 
+import Control.Controller;
+
+import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.Date;
 
 public class History {
     private String logBarcode;
-    private Date date;
+    private LocalDateTime date;
     private double cost;
     private double offSCost;
-    private ArrayList<Product> products;
-    private String name;
+    private ArrayList<String> products;
+    private ArrayList<String> name;
     private boolean deliveredOrNot;
 
-    public History(String logBarcode, Date date, double paidCost, double offSCost, String sellerName, boolean deliveredOrNot) {
-        this.logBarcode = logBarcode;
+    public History(LocalDateTime date, double paidCost, double offSCost, ArrayList<String> sellerName, ArrayList<String> products) {
+        this.logBarcode = Product.givenUsingPlainJava_whenGeneratingRandomStringUnbounded_thenCorrect();
         this.date = date;
         this.cost = paidCost;
         this.offSCost = offSCost;
-        this.products = new ArrayList<Product>();
+        this.products = products;
         this.name = sellerName;
-        this.deliveredOrNot = deliveredOrNot;
+    }
+
+    public String getLogBarcode() {
+        return logBarcode;
     }
 }
 
