@@ -1,31 +1,34 @@
 package Model;
 
+
+
 import java.io.File;
 import java.util.ArrayList;
 
 public class Manager extends Account {
 
-    private static boolean isFirstManagerCreatedOrNot;
+    private static boolean isFirstManagerCreatedOrNot = false;
 
-    private static ArrayList<Manager> allManagers;
-    private static ArrayList<Manager> managers;
 
-    private static ArrayList<RequestANewSellerAccount> registerSellerAccountRequests;
-    private static ArrayList<RequestProduct> editProductsRequests;
-    private static ArrayList<RequestOff> editOffRequests;
-    private static ArrayList<Account> allAccounts;
-    private static ArrayList<Category> categories;
+    private static ArrayList<RequestANewSellerAccount> registerSellerAccountRequests = new ArrayList<>();
+    private static ArrayList<RequestProduct> editProductsRequests = new ArrayList<>();
+    private static ArrayList<RequestOff> editOffRequests = new ArrayList<>();
+    private static ArrayList<SaveAble> allAccounts = new ArrayList<>();
+    private static final ArrayList<SaveAble> categories = new ArrayList<>();
 
-    private static ArrayList<CodedOff> offCodes;
+
 
 
     public Manager(String userName, String passWord) {
         super(userName, passWord);
-        allManagers.add(this);
-        isFirstManagerCreatedOrNot = false;
+        SaveAndLoad.getSaveAndLoad().writeJSON(this, Manager.class, userName);
     }
 
     public static void removeDiscount(String offName) {
+        if (CodedOff.getAllDiscounts().contains(CodedOff.getOffCodeWithName(offName))) {
+            CodedOff.getAllDiscounts().remove(CodedOff.getOffCodeWithName(offName));
+        }
+        SaveAndLoad.getSaveAndLoad().writeJSON(CodedOff.getAllDiscounts(), ArrayList.class, "offCodes");
     }
 
     public void editOffCode(CodedOff offCode) {
@@ -36,16 +39,17 @@ public class Manager extends Account {
         return registerSellerAccountRequests;
     }
 
-    public static ArrayList<RequestProduct> getEditProductsRequest() {
+    public static ArrayList<RequestProduct> getEditProductsRequests() {
         return editProductsRequests;
     }
+
 
     public static ArrayList<RequestOff> getEditOffRequests() {
         return editOffRequests;
     }
 
-    public static ArrayList<Request> getAllRequests() {
-        ArrayList<Request> allRequests = new ArrayList<>();
+    public static ArrayList<SaveAble> getAllRequests() {
+        ArrayList<SaveAble> allRequests = new ArrayList<>();
         allRequests.addAll(registerSellerAccountRequests);
         allRequests.addAll(editOffRequests);
         allRequests.addAll(editProductsRequests);
@@ -56,25 +60,34 @@ public class Manager extends Account {
 
     }
 
-    public static void remove(Product productWithName) {
-
-    }
-
     public static boolean addANewManager(String userName, String passWord, Boolean isRequestFromManger) {
-        if (isFirstManagerCreatedOrNot && isRequestFromManger){
+        File directory = new File(System.getProperty("user.dir") + "\\" + Manager.class);
+        if (directory.isDirectory()) {
+            String[] files = directory.list();
+            if (files.length > 0) {
+                System.out.println(System.getProperty("user.dir") + "\\" + Manager.class);
+                isFirstManagerCreatedOrNot = true;
+            }
+            else {
+                isFirstManagerCreatedOrNot = false;
+            }
+        }
+        if ((!isFirstManagerCreatedOrNot) || isRequestFromManger){
+            //System.out.println(System.getProperty("user.dir") + "\\" + Manager.class);
             new Manager(userName, passWord);
+            isFirstManagerCreatedOrNot = false;
             return true;
         }
         return false;
     }
 
-    public static boolean addANewSellerRequest(String userName, String passWord, String requestId) {
-        registerSellerAccountRequests.add(new RequestANewSellerAccount(requestId, "Create a seller account", userName, passWord));
-        return true;
-    }
+//    public static boolean addANewSellerRequest(String userName, String passWord, String requestId) {
+//        registerSellerAccountRequests.add(new RequestANewSellerAccount(requestId, "Create a seller account", userName, passWord));
+//        return true;
+//    }
 
 
-    public ArrayList<Account> getAllAccounts() {
+    public ArrayList<SaveAble> getAllAccounts() {
         return allAccounts;
     }
 
@@ -94,19 +107,6 @@ public class Manager extends Account {
 
     }
 
-    public static void removeProduct(Product product) {
-        File file = new File(product.getName());
-
-        if(file.delete())
-        {
-            System.out.println("File deleted successfully");
-        }
-        else
-        {
-            System.out.println("Failed to delete the file");
-        }
-        Product.removeProduct(product);
-    }
 
     public static void AnswerRequest(Request request, boolean acceptOrDecline) {
 
@@ -128,16 +128,29 @@ public class Manager extends Account {
 
     private static boolean accountRequestAccept(RequestANewSellerAccount request) {
         if (registerSellerAccountRequests.contains(request)) {
-            SaveAndLoad.getSaveAndLoad().writeJSONAccount(new Customer(request.getUserName(), request.getPassWord()));
+            new Seller(request.getUserName(), request.getPassWord());
+            registerSellerAccountRequests.remove(request);
+            SaveAndLoad.getSaveAndLoad().writeJSON(registerSellerAccountRequests, ArrayList.class, "registerSellerAccountRequests");
             return true;
         }
         return false;
     }
 
+    public static boolean addANewSellerRequest(String userName, String passWord) {
+        registerSellerAccountRequests.add(new RequestANewSellerAccount("Create a seller account", userName, passWord));
+        SaveAndLoad.getSaveAndLoad().writeJSON(registerSellerAccountRequests, ArrayList.class, "registerSellerAccountRequests");
+        return true;
+    }
+
     private static boolean editProduct(RequestProduct request) {
         if (editProductsRequests.contains(request)) {
+            Product product = Product.getProductWithName(request.getProductName());
+            if (Product.getAllProducts().contains(product)) {
+                Product.getAllProducts().remove(product);
+            }
             new Product(request.getProduct());
             editProductsRequests.remove(request);
+            SaveAndLoad.getSaveAndLoad().writeJSON(editOffRequests, ArrayList.class, "editOffRequests");
             return true;
         }
         return false;
@@ -145,8 +158,13 @@ public class Manager extends Account {
 
     private static boolean editOff(RequestOff request) {
         if (editOffRequests.contains(request)) {
+            Off off = Off.getOffByName(request.getOffName());
+            if (Off.getAllOffs().contains(off)) {
+                Off.getAllOffs().remove(off);
+            }
             new Off(request.getOff().getOffBarcode(), request.getOff().getStartDate(), request.getOff().getEndDate(), request.getOff().getOffAmount());
             editOffRequests.remove(request);
+            SaveAndLoad.getSaveAndLoad().writeJSON(editOffRequests, ArrayList.class, "editOffRequests");
             return true;
         }
         return false;
@@ -167,7 +185,7 @@ public class Manager extends Account {
     }
 
     public static void setRegisterSellerAccountRequest(Request registerSellerAccountRequests) {
-        Manager.registerSellerAccountRequests.add((RequestANewSellerAccount) registerSellerAccountRequests);
+        //Manager.registerSellerAccountRequests.add((RequestANewSellerAccount) registerSellerAccountRequests);
     }
 
     public static void setEditProductsRequest(Request editProductsRequests) {
@@ -177,6 +195,27 @@ public class Manager extends Account {
     public static void setEditOffRequests(Request editOffRequests) {
         Manager.editOffRequests.add((RequestOff) editOffRequests);
     }
+
+    public static Request getRequestByName(String id) {
+        for (Request request : registerSellerAccountRequests) {
+            if (request.getRequestId().equalsIgnoreCase(id)) {
+                return request;
+            }
+        }
+        for (Request request : editOffRequests) {
+            if (request.getRequestId().equalsIgnoreCase(id)) {
+                return request;
+            }
+        }
+        for (Request request : editProductsRequests) {
+            if (request.getRequestId().equalsIgnoreCase(id)) {
+                return request;
+            }
+        }
+        return null;
+    }
+
+
 
     @Override
     public String toString() {
