@@ -3,14 +3,13 @@ package Model;
 import java.util.ArrayList;
 
 public class RequestANewSellerAccount extends Request {
-    private static ArrayList<RequestANewSellerAccount> allRequestANewSellerAccounts;
+    private static ArrayList<RequestANewSellerAccount> allRequestANewSellerAccounts = new ArrayList<>();
     private String userName;
     private String passWord;
-    public RequestANewSellerAccount(String requestId, String requestType, String userName, String passWord) {
-        super(requestType, requestId);
+    public RequestANewSellerAccount(String requestType, String userName, String passWord) {
+        super(requestType);
         this.userName = userName;
         this.passWord = passWord;
-        allRequestANewSellerAccounts.add(this);
     }
 
     public String getUserName() {
@@ -19,5 +18,14 @@ public class RequestANewSellerAccount extends Request {
 
     public String getPassWord() {
         return passWord;
+    }
+
+    @Override
+    public String toString() {
+        return "RequestANewSellerAccount{" +
+                "userName='" + userName + '\'' +
+                ", passWord='" + passWord + '\'' +
+                ", requestId=" + requestId +
+                '}';
     }
 }
