@@ -6,26 +6,22 @@ public class Seller extends Account {
 
     //private static ArrayList<Seller> allSellers = new ArrayList<Seller>();
     protected ArrayList<History> history;
-    private ArrayList<Product> products;
+    private ArrayList<String> products;
     protected String companyName;
 
     private ArrayList<History> sellHistory;
-    private ArrayList<Product> sellingProducts;
-    private ArrayList<Off> offs;
+    private ArrayList<String> sellingProducts;
+    private ArrayList<String> offs;
 
     public Seller(String userName, String passWord) {
         super(userName, passWord);
         this.sellHistory = new ArrayList<History>();
-        this.sellingProducts = new ArrayList<Product>();
-        this.offs = new ArrayList<Off>();
+        this.sellingProducts = new ArrayList<>();
+        this.offs = new ArrayList<>();
         this.history = new ArrayList<>();
         this.products = new ArrayList<>();
+        SaveAndLoad.getSaveAndLoad().writeJSON(this, Seller.class, userName);
         //allSellers.add(this);
-    }
-
-    public static boolean addANewSeller(String id, String userName, String passWord) {
-        Manager.addANewSellerRequest(id, userName, passWord);
-        return true;
     }
 
     public void setCompanyName(String companyName) {
@@ -36,7 +32,7 @@ public class Seller extends Account {
         this.sellHistory.add(soldProduct);
     }
 
-    public void setSellingProducts(ArrayList<Product> sellingProducts) {
+    public void setSellingProducts(ArrayList<String> sellingProducts) {
         this.sellingProducts = sellingProducts;
     }
 
@@ -58,6 +54,22 @@ public class Seller extends Account {
 
     public void sendEditOffRequest() {
         //Manager.receiveEditOffRequest();
+    }
+
+    public ArrayList<String> getOffs() {
+        return offs;
+    }
+
+    public ArrayList<History> getHistory() {
+        return history;
+    }
+
+    public ArrayList<String> getProducts() {
+        return products;
+    }
+
+    public String getCompanyName() {
+        return companyName;
     }
 
     @Override
