@@ -1,7 +1,6 @@
 package View.Menu;
 
 import Control.Controller;
-import View.CommandsSource;
 
 import static View.Outputs.*;
 
@@ -13,36 +12,33 @@ public class OffsMenu extends Menu {
         options.add("show product [productId]");
         options.add("help");
         options.add("back");
-        options.add("logout");
     }
 
     public static void offsList() {
         printOffsListResult(Controller.getOurController().requestOffsList());
     }
 
-    private void showProduct(String[] splitInput, String input) {
-        productMenu.setProductId(splitInput[2]);
-        productMenu.execute();
-    }
-
     public void execute() {
         offsList();
-        System.out.println("Enter number :");
-        while(true) {
-            showCommands();
-            String input;
-            if(!getMatcher(input = scanner.nextLine().trim(), "(\\d)").matches()){
+        String input = "";
+        do {
+            show();
+            System.out.println("Enter Number :");
+            if (!isThisRegexMatch("(\\d)", input = scanner.nextLine())) {
                 continue;
             }
-            switch (input) {
+            switch (input.trim()) {
                 case "1":
-                    System.out.println("show product");
+                    String productId = getField("enter productId", "(\\S+)").group(1);
+                    ProductMenu productMenu = new ProductMenu();
+                    productMenu.execute(productId);
                     break;
                 case "2":
+                    show();
+                    break;
+                case "3":
                     return;
-                default:
-                    DefaultMenu.getInstance().execute(Integer.parseInt(input) - 2);
             }
-        }
+        }while (true);
     }
 }
